% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind.fake.data.frame.R
\name{cbind.fake.data.frame}
\alias{cbind.fake.data.frame}
\title{Binds a fake.data.frame to another fake.data.frame or a data.frame.}
\usage{
\method{cbind}{fake.data.frame}(x, ...)
}
\arguments{
\item{x}{a fake.data.frame}
}
\description{
Binds a fake.data.frame to another fake.data.frame or a \code{data.frame}.
}

